manifest: hal-ti: HwiP: added AUX software event interrupts
Changed hal_ti revision to represent the pull request 52

Added AUX_SWEV0 and AUX_SWEV1 interrupts to allow the Sensor Controller to
trigger these events and the System CPU to evaluate them. Also changed 
HwiP_Params_init and the enabling of interrupts in HwiP_construct to allow
interrupts to be constructed without directly enabling them. This is used
by the sensor controller interface when registering the interrupts.

Signed-off-by: Benedikt Streicher <streicher.b@posteo.de>